+++ pointcept/datasets/utils.py
@@ -14,19 +14,6 @@ from torch.utils.data.dataloader import default_collate
 import sys
 import pdb
 
-class ForkedPdb(pdb.Pdb):
-    """A Pdb subclass that may be used
-    from a forked multiprocessing child
-
-    """
-    def interaction(self, *args, **kwargs):
-        _stdin = sys.stdin
-        try:
-            sys.stdin = open('/dev/stdin')
-            pdb.Pdb.interaction(self, *args, **kwargs)
-        finally:
-            sys.stdin = _stdin
-
 def collate_fn(batch):
     """
     collate function for point cloud which support dict and list,
